requirements.txt
fastapi
uvicorn
pyfiglet
scikit-learn == 1.4.2
nltk
python-dotenv
boto3
celery[redis]
redis
prometheus-client

